Remove Prometheus service
