+++ app/controllers/concerns/collect_data_dog_metrics.rb
@@ -1,0 +1,37 @@
+# frozen_string_literal: true
+
+module CollectDataDogMetrics
+  extend ActiveSupport::Concern
+
+  included do
+    before_action :collect_data_dog_metrics
+  end
+
+  def collect_data_dog_metrics
+    collect_postgres_metrics
+  end
+
+  def collect_postgres_metrics
+    conns = ActiveRecord::Base.connection_pool.connections
+
+    active = conns.count { |c| c.in_use? && c.owner.alive? }
+    dead = conns.count { |c| c.in_use? && !c.owner.alive? }
+    idle = conns.count { |c| !c.in_use? }
+
+    emit_datadog_point("postgres", "active", active)
+    emit_datadog_point("postgres", "dead", dead)
+    emit_datadog_point("postgres", "idle", idle)
+  end
+
+  def emit_datadog_point(db_name, type, count)
+    DataDogService.emit_gauge(
+      metric_group: "database",
+      metric_name: "#{type}_connections",
+      metric_value: count,
+      app_name: "caseflow",
+      attrs: {
+        database: db_name
+      }
+    )
+  end
+end
