+++ app/controllers/health_checks_controller.rb
@@ -1,4 +1,5 @@
 class HealthChecksController < ApplicationController
+  include CollectDataDogMetrics
   skip_before_action :authenticate
   skip_before_action :check_out_of_service
   skip_before_action :check_v2_app_access
