+++ config/secrets.yml
@@ -19,7 +19,6 @@ development:
   secret_key_base: 8c6b64e47fa8d523a9ea3d7e7658dbd5ee916003785039b14ffdd076fae644afaedc04eb683fc705c47f6f07020184c8cf3a30e79f2258521a45b861ab8e7862
   redis_url_cache: <%= ENV["REDIS_URL_CACHE"] || "redis://localhost:16379/0/cache/" %>
   redis_url: <%= ENV["REDIS_URL_SIDEKIQ"] || "redis://localhost:16379" %>
-  prometheus_push_gateway_url: <%= ENV["PROMETHEUS_PUSH_GATEWAY_URL"] %>
   vbms:
     url: https://filenet.uat.vbms.aide.oit.va.gov/vbmsp2-cms/streaming/eDocumentService-v4
     env_dir: ../deployment/decrypted_creds/uat_creds
@@ -34,7 +33,6 @@ staging:
   secret_key_base: 8c6b64e47fa8d523a9ea3d7e7658dbd5ee916003785039b14ffdd076fae644afaedc04eb683fc705c47f6f07020184c8cf3a30e79f2258521a45b861ab8e7862
   redis_url_cache: <%= ENV["REDIS_URL_CACHE"] || "redis://localhost:6379/0/cache/" %>
   redis_url: <%= ENV["REDIS_URL_SIDEKIQ"] || "redis://localhost:6379" %>
-  prometheus_push_gateway_url: <%= ENV["PROMETHEUS_PUSH_GATEWAY_URL"] %>
   vbms:
     env: uat
     url: https://filenet.uat.vbms.aide.oit.va.gov/vbmsp2-cms/streaming/eDocumentService-v4
@@ -50,14 +48,12 @@ demo:
   secret_key_base: 8c6b64e47fa8d523a9ea3d7e7658dbd5ee916003785039b14ffdd076fae644afaedc04eb683fc705c47f6f07020184c8cf3a30e79f2258521a45b861ab8e7862
   redis_url_cache: <%= ENV["REDIS_URL_CACHE"] || "redis://localhost:6379/0/cache/" %>
   redis_url: <%= ENV["REDIS_URL_SIDEKIQ"] || "redis://localhost:6379" %>
-  prometheus_push_gateway_url: <%= ENV["PROMETHEUS_PUSH_GATEWAY_URL"] %>
 
 
 test:
   secret_key_base: 35cf2535cc790c92a07bcab48a5f21bbd132a7071984947e8cc92d2ce8e60d582c62022b7ddcfa6f0eca34319dfe1482fbf21d2c78e50c12e884be9ea5fea7e0
   redis_url_cache: <%= ENV["REDIS_URL_CACHE"] || "redis://localhost:6379/0/cache/" %>
   redis_url: <%= ENV["REDIS_URL_SIDEKIQ"] || "redis://localhost:6379" %>
-  prometheus_push_gateway_url: <%= ENV["PROMETHEUS_PUSH_GATEWAY_URL"] %>
 
 # Do not keep production secrets in the unencrypted secrets file.
 # Instead, either read values from the environment.
@@ -67,6 +63,5 @@ production:
   secret_key_base: <%= ENV["SECRET_KEY_BASE"] %>
   redis_url_cache: <%= ENV["REDIS_URL_CACHE"] %>
   redis_url: <%= ENV["REDIS_URL_SIDEKIQ"] %>
-  prometheus_push_gateway_url: <%= ENV["PROMETHEUS_PUSH_GATEWAY_URL"] %>
   vbms:
       env: true
